Rename Compass service to Trace
