--- internal/server/mock/compass/server.go
@@ -1,10 +1,0 @@
-package compass
-
-import (
-	"github.com/skpr/api/pb"
-)
-
-// Server implements the GRPC "compass" definition.
-type Server struct {
-	pb.UnimplementedCompassServer
-}
