+++ internal/server/mock/trace/server.go
@@ -1,0 +1,10 @@
+package trace
+
+import (
+	"github.com/skpr/api/pb"
+)
+
+// Server implements the GRPC "trace" definition.
+type Server struct {
+	pb.UnimplementedTraceServer
+}
